refactor: combine the two definitions of `create_test_nodes` into one (#171)
We had a similar definition in two places: `src/main.rs` and
`tests/liveness.rs`.

Unfortunately, the shared code has to go into `src/lib.rs`.
