+++ src/lib.rs
@@ -19,6 +19,7 @@ pub mod types;
 pub mod validator;
 
 use std::net::SocketAddr;
+use std::sync::Arc;
 
 use serde::{Deserialize, Serialize};
 
@@ -29,6 +30,12 @@ use self::crypto::{Hash, aggsig, signature};
 pub use self::disseminator::Disseminator;
 use self::types::Slot;
 pub use self::validator::Validator;
+use crate::all2all::TrivialAll2All;
+use crate::consensus::EpochInfo;
+use crate::crypto::signature::SecretKey;
+use crate::disseminator::Rotor;
+use crate::disseminator::rotor::StakeWeightedSampler;
+use crate::network::{UdpNetwork, localhost_ip_sockaddr};
 
 /// Validator ID number type.
 pub type ValidatorId = u64;
@@ -78,4 +85,78 @@ pub(crate) fn highest_non_zero_byte(mut val: usize) -> usize {
         cnt += 1;
     }
     cnt
+}
+
+type TestNode =
+    Alpenglow<TrivialAll2All<UdpNetwork>, Rotor<UdpNetwork, StakeWeightedSampler>, UdpNetwork>;
+
+struct Networks {
+    all2all: UdpNetwork,
+    disseminator: UdpNetwork,
+    repair: UdpNetwork,
+    txs: UdpNetwork,
+}
+
+impl Networks {
+    fn new() -> Self {
+        Self {
+            all2all: UdpNetwork::new_with_any_port(),
+            disseminator: UdpNetwork::new_with_any_port(),
+            repair: UdpNetwork::new_with_any_port(),
+            txs: UdpNetwork::new_with_any_port(),
+        }
+    }
+}
+
+/// Creates [`TestNode`] for testing and benchmarking purposes.
+///
+/// This code lives here to enable sharing between different testing and benchmarking.
+/// It should not be used in production code.
+pub fn create_test_nodes(count: u64) -> Vec<TestNode> {
+    // open sockets with arbitrary ports
+    let networks = (0..count).map(|_| Networks::new()).collect::<Vec<_>>();
+
+    // prepare validator info for all nodes
+    let mut rng = rand::rng();
+    let mut sks = Vec::new();
+    let mut voting_sks = Vec::new();
+    let mut validators = Vec::new();
+    for (id, network) in networks.iter().enumerate() {
+        sks.push(SecretKey::new(&mut rng));
+        voting_sks.push(aggsig::SecretKey::new(&mut rng));
+        let all2all_address = localhost_ip_sockaddr(network.all2all.port());
+        let disseminator_address = localhost_ip_sockaddr(network.disseminator.port());
+        let repair_address = localhost_ip_sockaddr(network.repair.port());
+        validators.push(ValidatorInfo {
+            id: id as u64,
+            stake: 1,
+            pubkey: sks[id].to_pk(),
+            voting_pubkey: voting_sks[id].to_pk(),
+            all2all_address,
+            disseminator_address,
+            repair_address,
+        });
+    }
+
+    // turn validator info into actual nodes
+    networks
+        .into_iter()
+        .enumerate()
+        .map(|(id, network)| {
+            let epoch_info = Arc::new(EpochInfo::new(id as u64, validators.clone()));
+            let all2all = TrivialAll2All::new(validators.clone(), network.all2all);
+            let disseminator = Rotor::new(network.disseminator, epoch_info.clone());
+            let repair_network = network.repair;
+            let txs_receiver = network.txs;
+            Alpenglow::new(
+                sks[id].clone(),
+                voting_sks[id].clone(),
+                all2all,
+                disseminator,
+                repair_network,
+                epoch_info,
+                txs_receiver,
+            )
+        })
+        .collect()
 }
